requirements.txt
numpy>=1.21
h5py>=3.1
Pillow>=9.0
matplotlib>=3.4
scipy>=1.7
streamlit>=1.28
